feat: refactor generate-signature && add unit test
